Fix remaining warnings on MSVC.

+++ examples/cpp-config-ui.cpp
@@ -212,10 +212,10 @@ int main(int argc, char * argv[]) try
             y += 38;
         }
         g.label({w-260,y+12}, {1,1,1}, "Depth control parameters preset: %g", dc_preset);
-        if(g.slider(100, {w-260,y+16,w-20,y+36}, 0, 5, 1, dc_preset)) rs_apply_depth_control_preset((rs_device *)dev, dc_preset);
+        if(g.slider(100, {w-260,y+16,w-20,y+36}, 0, 5, 1, dc_preset)) rs_apply_depth_control_preset((rs_device *)dev, static_cast<int>(dc_preset));
         y += 38;
         g.label({w-260,y+12}, {1,1,1}, "IVCAM options preset: %g", iv_preset);
-        if(g.slider(101, {w-260,y+16,w-20,y+36}, 0, 8, 1, iv_preset)) rs_apply_ivcam_preset((rs_device *)dev, iv_preset);
+        if(g.slider(101, {w-260,y+16,w-20,y+36}, 0, 8, 1, iv_preset)) rs_apply_ivcam_preset((rs_device *)dev, static_cast<int>(iv_preset));
         y += 38;
         
         panel_height = y + 10 + offset;
